update mobile view and move components folder
